Move TextInput PropTypes to Deprecated PropTypes
